django/localomd/requirements.txt
Django==1.10.3
django-crispy-forms==1.6.0
django-pagedown==0.1.1
djangorestframework==3.4.7
django-cors-headers==1.3.1
djangorestframework-jwt==1.8.0
mod-wsgi==4.5.7
mod-wsgi-httpd==2.4.23.1
pytz==2016.7
requests==2.12.4

